requirements.txt
# Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6

# LLM + embeddings
openai==1.57.0
anthropic==0.39.0

# File parsing
pypdf==5.1.0
python-docx==1.1.2

# Auth
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
email-validator==2.2.0

